Added collision letter animation called 'LetterGone' (made to sprite).
This animation will play when the human and rat characters collide with the letters in their corresponding areas - a fade transition will begin as soon as the animation plays - this will fade into the cutscene areas. For example, when the human character collides with the letter object in Area1 -the fade and collide animation will trigger - when cutscene finishes, fades to black and back to Area1 and the oHumanToRat animation will play then the rat level in Area2 can begin.